Increased bit width of set length/capacity values

+++ include/scene.h
@@ -17,7 +17,7 @@ typedef struct {
 
 void ConstructScene(Scene* s, Camera c, Light);
 void DeconstructScene(Scene* s);
-unsigned AddShape(Scene* s, Shape sp);
+unsigned long AddShape(Scene* s, Shape sp);
 void IntersectScene(Scene* s, Ray r, Set* intersection_set);
 void RenderScene(Scene* s, Canvas* c);
 bool IsInShadow(Scene *s, Tuple3 location);
