+++ include/set.h
@@ -3,15 +3,15 @@
 
 typedef struct {
     void* data;
-    unsigned length;
-    unsigned capacity;
-    unsigned data_width;
+    unsigned long length;
+    unsigned long capacity;
+    unsigned long data_width;
 } Set;
 
 void ConstructSet(Set* s, unsigned data_width);
 void DeconstructSet(Set* s);
-unsigned AppendValue(Set* s, void* value);
-void* Index(Set* s, unsigned index);
-void CopyOut(Set* s, unsigned index, void* out);
+unsigned long AppendValue(Set* s, void* value);
+void* Index(Set* s, unsigned long index);
+void CopyOut(Set* s, unsigned long index, void* out);
 
 #endif
